requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.35.0
google-genai==1.67.0
python-dotenv==1.0.0
websockets==15.0.1

# Testing
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-timeout==2.3.1
httpx==0.28.1

# Harness
pyyaml==6.0.2
Pillow==11.2.1
